requirements.txt
pyserial~=3.5
requests~=2.28.1
pytz~=2022.1